requirements.txt
pybamm>=24.3
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
scipy>=1.10
statsmodels>=0.14
pyyaml>=6.0
